+++ NEWS.md
@@ -40,6 +40,11 @@
   sections, R examples, Python examples, and SQL examples for the new
   operationalization workflow and renamed arguments.
 
+## Bug Fixes
+
+* Use `get()` instead of `quote()` so that the methods will work when the
+  namespace is load and not attached. (#20)
+
 # Version 1.1.3:
 
 * Spelling and grammar fixes in the documentation for the R, python, and